move prototype to controller
